将热门推荐、景点列表、城市列表的Activity改为Fragment，以适配Drawer的切换。
