--- listviewtest/src/main/java/com/jason/listviewtest/activity/SpotListFragment.java
@@ -1,7 +1,0 @@
-package com.jason.listviewtest.activity;
-
-/**
- * Created by Jason on 2016/7/22.
- */
-public class SpotListFragment {
-}
