+++ listviewtest/src/main/java/com/jason/listviewtest/fragment/CityListFragment.java
@@ -1,10 +1,11 @@
-package com.jason.listviewtest.activity;
+package com.jason.listviewtest.fragment;
+
 
 import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.support.v4.app.Fragment;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -13,9 +14,10 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.jason.listviewtest.adapter.CityListAdapter;
-import com.jason.listviewtest.R;
 import com.jason.listviewtest.Helpter.Utils;
+import com.jason.listviewtest.R;
+import com.jason.listviewtest.activity.CityDetailActivity;
+import com.jason.listviewtest.adapter.CityListAdapter;
 import com.jason.listviewtest.imageloader.ImageLoader;
 import com.jason.listviewtest.model.City;
 import com.jason.listviewtest.model.RecyclerViewItemClickListener;
@@ -25,30 +27,77 @@ import com.timehop.stickyheadersrecyclerview.StickyRecyclerHeadersDecoration;
 
 import java.util.Map;
 
-public class CityListActivity extends AppCompatActivity {
+/**
+ * A simple {@link Fragment} subclass.
+ * Use the {@link CityListFragment#newInstance} factory method to
+ * create an instance of this fragment.
+ */
+public class CityListFragment extends Fragment {
+    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
+    private static final String ARG_PARAM1 = "param1";
+    private static final String ARG_PARAM2 = "param2";
+
+    private String mParam1;
+    private String mParam2;
+
 
     private RecyclerView mRecyclerView;
     private ImageLoader imageLoader;
 
     private Bitmap defaultRoundBitmap;
+    
+    private View mFragView;
 
+    public CityListFragment() {
+        // Required empty public constructor
+    }
+
+    /**
+     * Use this factory method to create a new instance of
+     * this fragment using the provided parameters.
+     *
+     * @param param1 Parameter 1.
+     * @param param2 Parameter 2.
+     * @return A new instance of fragment CityListFragment.
+     */
+    public static CityListFragment newInstance(String param1, String param2) {
+        CityListFragment fragment = new CityListFragment();
+        Bundle args = new Bundle();
+        args.putString(ARG_PARAM1, param1);
+        args.putString(ARG_PARAM2, param2);
+        fragment.setArguments(args);
+        return fragment;
+    }
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        if (getArguments() != null) {
+            mParam1 = getArguments().getString(ARG_PARAM1);
+            mParam2 = getArguments().getString(ARG_PARAM2);
+        }
+    }
 
-        mRecyclerView = (RecyclerView) findViewById(R.id.recyclerView);
-        mRecyclerView.setLayoutManager(new LinearLayoutManager(this,LinearLayoutManager.VERTICAL, false));
-
-        Utils.initProvinceMap();
-        Utils.initCityList(getApplicationContext());
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        mFragView = inflater.inflate(R.layout.activity_city_list, container, false);
+        mRecyclerView = (RecyclerView) mFragView.findViewById(R.id.recyclerView);
+        mRecyclerView.setLayoutManager(new LinearLayoutManager(getContext(),LinearLayoutManager.VERTICAL, false));
+        
         //init imageLoader
-        imageLoader = ImageLoader.build(CityListActivity.this);
+        imageLoader = ImageLoader.build(getContext());
 
         //set default bitmap in the list
         defaultRoundBitmap = Utils.toRoundBitmap(BitmapFactory.decodeResource(getResources(), R.drawable.image_default));
 
+        initRecyclerView();
+        
+        return mFragView;
+    }
+
+    private void initRecyclerView() {
         CityListWithHeaderAdapter adapter = new CityListWithHeaderAdapter();
 
 
@@ -60,7 +109,7 @@ public class CityListActivity extends AppCompatActivity {
             public void onItemClick(View view, int position) {
                 City city = Utils.listCity.get(position);
                 if(city != null) {
-                    Intent i = new Intent(CityListActivity.this,CityDetailActivity.class);
+                    Intent i = new Intent(getContext(),CityDetailActivity.class);
                     i.putExtra("CityPos",position);
                     startActivity(i);
                 }
@@ -68,11 +117,9 @@ public class CityListActivity extends AppCompatActivity {
         });
         mRecyclerView.setAdapter(adapter);
 
-
         //set header
         final StickyRecyclerHeadersDecoration headersDecor = new StickyRecyclerHeadersDecoration(adapter);
         mRecyclerView.addItemDecoration(headersDecor);
-
     }
 
     /**
