+++ listviewtest/src/main/res/layout/activity_city_list.xml
@@ -1,13 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingLeft="8dp"
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    android:paddingRight="8dp"
-    tools:context=".activity.CityListActivity">
+    android:paddingRight="8dp">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recyclerView"
