+++ listviewtest/src/main/res/layout/activity_spot_list.xml
@@ -1,13 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="8dp"
     android:paddingRight="8dp"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.jason.listviewtest.activity.SpotListActivity">
+    android:paddingTop="@dimen/activity_vertical_margin">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recyclerView_spot"
