+++ listviewtest/src/main/res/layout/app_bar_main_page.xml
@@ -21,7 +21,7 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <include layout="@layout/content_main_page" />
+    <include layout="@layout/default_main_content" />
 
     <android.support.design.widget.FloatingActionButton
         android:id="@+id/fab"
