+++ listviewtest/src/main/res/menu/activity_main_page_drawer.xml
@@ -3,6 +3,10 @@
 
     <group android:checkableBehavior="single">
         <item
+            android:id="@+id/nav_recommend"
+            android:icon="@drawable/ic_menu_camera"
+            android:title="热门推荐" />
+        <item
             android:id="@+id/nav_city_list"
             android:icon="@drawable/ic_menu_gallery"
             android:title="城市列表" />
@@ -10,10 +14,6 @@
             android:id="@+id/nav_spot_list"
             android:icon="@drawable/ic_menu_gallery"
             android:title="景点列表" />
-        <item
-            android:id="@+id/nav_footprint"
-            android:icon="@drawable/ic_menu_camera"
-            android:title="足迹" />
         <item
             android:id="@+id/nav_query"
             android:icon="@drawable/ic_menu_share"
